feat: attempt refactoring to Illuminate Manager
Closes #3
