+++ composer.json
@@ -6,6 +6,9 @@
     "require": {
         "php": "^8.1",
         "guzzlehttp/guzzle": "^7.4",
+        "illuminate/config": "^9.7",
+        "illuminate/container": "^9.7",
+        "illuminate/support": "^8.0 || ^9.0",
         "spatie/data-transfer-object": "^3.7"
     },
     "require-dev": {
