+++ src/Contracts/Driver.php
@@ -7,9 +7,6 @@ use ParcelTrap\DTOs\TrackingDetails;
 
 interface Driver
 {
-    /** @param array<string, mixed> $config */
-    public static function make(array $config, ?ClientInterface $client = null): self;
-
     /** @param  array<string, mixed>  $parameters */
     public function find(string $identifier, array $parameters = []): TrackingDetails;
 }
