+++ src/Contracts/Factory.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ParcelTrap\Contracts;
+
+interface Factory
+{
+    /**
+     * Get a ParcelTrap driver implementation.
+     *
+     * @param  string  $driver
+     * @return Driver
+     */
+    public function driver($driver = null);
+}
