+++ src/Drivers/NullDriver.php
@@ -1,20 +1,14 @@
 <?php
 
-namespace ParcelTrap\Tests;
+namespace ParcelTrap\Drivers;
 
 use DateTimeImmutable;
-use GuzzleHttp\ClientInterface;
 use ParcelTrap\Contracts\Driver;
 use ParcelTrap\DTOs\TrackingDetails;
 use ParcelTrap\Enums\Status;
 
 class NullDriver implements Driver
 {
-    public static function make(array $config, ?ClientInterface $client = null): Driver
-    {
-        return new self();
-    }
-
     public function find(string $identifier, array $parameters = []): TrackingDetails
     {
         return new TrackingDetails(
