+++ tests/ParcelTrapTest.php
@@ -2,26 +2,27 @@
 
 declare(strict_types=1);
 
+use ParcelTrap\Drivers\NullDriver;
 use ParcelTrap\DTOs\TrackingDetails;
 use ParcelTrap\Enums\Status;
 use ParcelTrap\ParcelTrap;
-use ParcelTrap\Tests\NullDriver;
 
 it('can instantiate ParcelTrap', function () {
-    expect(ParcelTrap::make(['null' => NullDriver::make([])]))
+    expect(ParcelTrap::make(['null' => new NullDriver()]))
         ->toBeInstanceOf(ParcelTrap::class);
 });
 
 it('can add multiple drivers to ParcelTrap', function () {
-    $client = ParcelTrap::make(['null' => NullDriver::make([])]);
-    $client->addDriver('null2', NullDriver::make([]));
+    $client = ParcelTrap::make(['null' => new NullDriver()]);
+    $client->extend('null2', function () {
+        return new NullDriver();
+    });
 
-    expect($client)->hasDriver('null2')->toBeTrue();
+    expect($client)->driver('null2')->toBeInstanceOf(NullDriver::class);
 });
 
 it('can retrieve a driver from ParcelTrap', function () {
-    expect(ParcelTrap::make(['null' => NullDriver::make([])]))
-        ->hasDriver('null')->toBeTrue()
+    expect(ParcelTrap::make(['null' => new NullDriver()]))
         ->driver('null')->toBeInstanceOf(NullDriver::class);
 });
 
@@ -30,23 +31,19 @@ it('throws an exception when a driver can\'t be found in ParcelTrap', function (
 })->throws(InvalidArgumentException::class);
 
 it('can set a default ParcelTrap driver', function () {
-    $client = ParcelTrap::make(['null' => NullDriver::make([])]);
+    $client = ParcelTrap::make(['null' => new NullDriver()]);
     $client->setDefaultDriver('null');
 
     expect($client->getDefaultDriver())->toBe('null');
     expect($client->driver())->toBeInstanceOf(NullDriver::class);
 });
 
-it('throws an exception when attempting to set a non-existant default driver in ParcelTrap', function () {
-    ParcelTrap::make([])->setDefaultDriver('null');
-})->throws(InvalidArgumentException::class);
-
 it('throws an exception when a default driver hasn\'t been set in ParcelTrap', function () {
     ParcelTrap::make([])->driver();
 })->throws(InvalidArgumentException::class);
 
 it('can call `find` on a ParcelTrap driver', function () {
-    expect(ParcelTrap::make(['null' => NullDriver::make([])])->driver('null')->find('abcdefg'))
+    expect(ParcelTrap::make(['null' => new NullDriver()])->driver('null')->find('abcdefg'))
         ->toBeInstanceOf(TrackingDetails::class)
         ->identifier->toBe('abcdefg')
         ->status->toBe(Status::Unknown)
@@ -56,7 +53,7 @@ it('can call `find` on a ParcelTrap driver', function () {
 });
 
 it('can call `find` on the default ParcelTrap driver', function () {
-    $client = ParcelTrap::make(['null' => NullDriver::make([])]);
+    $client = ParcelTrap::make(['null' => new NullDriver()]);
     $client->setDefaultDriver('null');
 
     expect($client->find('abcdefg'))->toBeInstanceOf(TrackingDetails::class);
